backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
Pillow==10.4.0
gunicorn==21.2.0
itsdangerous==2.2.0
Werkzeug==3.0.3
pytest==8.2.2
pytest-cov==5.0.0
requests==2.32.3

